go gennerate output
