+++ api/sk-serve/user/crossfunctional/models.go
@@ -1,0 +1,9 @@
+package crossfunctional
+
+type DecodedAccessCode struct {
+	LeagueID   string
+	LeagueName string
+	TeamID     string
+	Role       string
+	AccessCode string
+}
